Create module for Tachiyomi
