+++ kumaslash/src/main/java/me/ghostbear/kumaslash/commands/RulesEventHandler.java
@@ -12,7 +12,7 @@ import me.ghostbear.core.discord4j.annotations.DiscordInteractionProperties;
 import me.ghostbear.core.discord4j.annotations.DiscordComponent;
 import me.ghostbear.core.discord4j.annotations.DiscordInteractionHandler;
 import me.ghostbear.kumaslash.data.guild.GuildRuleRepository;
-import me.ghostbear.kumaslash.util.Tachiyomi;
+import me.ghostbear.kumaslash.tachiyomi.util.TachiyomiHelper;
 import reactor.core.publisher.Mono;
 
 import java.util.Optional;
@@ -49,7 +49,7 @@ public class RulesEventHandler {
 					.flatMap(ruleAndMention -> event.createFollowup()
 							.withContent(ruleAndMention.getT2())
 							.withEmbeds(EmbedCreateSpec.builder()
-									.color(Tachiyomi.color())
+									.color(TachiyomiHelper.color())
 									.title("Server Rule")
 									.description("%s. **%s** %s".formatted(ruleAndMention.getT1().index(), ruleAndMention.getT1().title(), ruleAndMention.getT1().description()))
 									.footer(EmbedCreateFields.Footer.of("Full list of rules can be found in #rules or by using /rules.", null))
@@ -72,7 +72,7 @@ public class RulesEventHandler {
 					.flatMap(rulesAndMention -> event.createFollowup()
 							.withContent(rulesAndMention.getT2())
 							.withEmbeds(EmbedCreateSpec.builder()
-									.color(Tachiyomi.color())
+									.color(TachiyomiHelper.color())
 									.title("Server Rules")
 									.description(rulesAndMention.getT1().trim())
 									.build()))
