--- kumaslash/src/main/java/me/ghostbear/kumaslash/data/tachiyomi/ExtensionRepository.java
@@ -1,6 +1,0 @@
-package me.ghostbear.kumaslash.data.tachiyomi;
-
-import org.springframework.data.repository.reactive.ReactiveCrudRepository;
-
-public interface ExtensionRepository extends ReactiveCrudRepository<Extension, String> {
-}
