--- kumaslash/src/main/java/me/ghostbear/kumaslash/data/tachiyomi/SourceMapper.java
@@ -1,18 +1,0 @@
-package me.ghostbear.kumaslash.data.tachiyomi;
-
-import me.ghostbear.kumaslash.data.tachiyomi.client.SourceDTO;
-
-public class SourceMapper {
-
-	public Source map(SourceDTO remote) {
-		return new Source(
-				remote.id(),
-				remote.name(),
-				remote.lang(),
-				remote.baseUrl(),
-				remote.versionId(),
-				remote.hasCloudflare() == 1
-		);
-	}
-
-}
