--- kumaslash/src/main/java/me/ghostbear/kumaslash/data/tachiyomi/client/SourceDTO.java
@@ -1,14 +1,0 @@
-package me.ghostbear.kumaslash.data.tachiyomi.client;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public record SourceDTO(
-		String name,
-		String lang,
-		String id,
-		String baseUrl,
-		int versionId,
-		int hasCloudflare
-) {
-}
