--- kumaslash/src/main/java/me/ghostbear/kumaslash/util/Tachiyomi.java
@@ -1,11 +1,0 @@
-package me.ghostbear.kumaslash.util;
-
-import discord4j.rest.util.Color;
-
-public interface Tachiyomi {
-
-	static Color color() {
-		return Color.of(44, 132, 194);
-	}
-
-}
