+++ service-tachiyomi/build.gradle
@@ -1,0 +1,37 @@
+plugins {
+    id "java"
+    id "org.springframework.boot" version "3.1.2"
+    id "io.spring.dependency-management" version "1.1.3"
+}
+
+springBoot {
+    buildInfo()
+}
+
+// Use a known archive name for Docker builds
+if (Objects.nonNull(System.getenv("DOCKER_PIPELINE")) && System.getenv("DOCKER_PIPELINE").toBoolean()) {
+    bootJar.configure {
+        archiveFileName.set("kumaslash.jar")
+    }
+}
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    api "com.discord4j:discord4j-core:3.2.5"
+    api "ch.qos.logback:logback-classic:1.4.8"
+
+    api "org.springframework.boot:spring-boot-starter-webflux"
+    api "org.springframework.boot:spring-boot-starter-data-r2dbc"
+
+    api "org.flywaydb:flyway-core"
+    api "org.apache.commons:commons-lang3"
+
+    runtimeOnly "org.postgresql:postgresql"
+    runtimeOnly "org.postgresql:r2dbc-postgresql"
+
+    api project(":core-discord4j")
+    api project(":service-github")
+}
