+++ service-tachiyomi/src/main/java/me/ghostbear/kumaslash/tachiyomi/configuration/TachiyomiConfiguration.java
@@ -1,0 +1,10 @@
+package me.ghostbear.kumaslash.tachiyomi.configuration;
+
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableConfigurationProperties(TachiyomiProperties.class)
+public class TachiyomiConfiguration {
+
+}
