+++ service-tachiyomi/src/main/java/me/ghostbear/kumaslash/tachiyomi/remote/model/SourceDTO.java
@@ -1,0 +1,14 @@
+package me.ghostbear.kumaslash.tachiyomi.remote.model;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public record SourceDTO(
+		String name,
+		String lang,
+		String id,
+		String baseUrl,
+		int versionId,
+		int hasCloudflare
+) {
+}
