+++ service-tachiyomi/src/main/java/me/ghostbear/kumaslash/tachiyomi/util/TachiyomiHelper.java
@@ -1,0 +1,11 @@
+package me.ghostbear.kumaslash.tachiyomi.util;
+
+import discord4j.rest.util.Color;
+
+public interface TachiyomiHelper {
+
+	static Color color() {
+		return Color.of(44, 132, 194);
+	}
+
+}
