+++ settings.gradle
@@ -3,6 +3,7 @@ rootProject.name = "kumaslash"
 include "core-discord4j"
 include "service-anilist"
 include "service-github"
+include "service-tachiyomi"
 include "kumaslash"
 include "kumaslash-anilist"
 include "samples:discord-bot"
